+++ server_test.go
@@ -1,14 +1,13 @@
 package graft
 
 import (
-	"github.com/benmills/quiz"
-	"github.com/wjdix/tiktok"
-	"testing"
 	"errors"
+	"github.com/benmills/quiz"
+	"testing"
 )
 
 type FailingPeer struct {
-	numberOfFails int
+	numberOfFails      int
 	successfulResponse VoteResponseMessage
 }
 
@@ -74,19 +73,6 @@ func TestLastLogTermDerivedFromLogEntries(t *testing.T) {
 	server.Log = []LogEntry{LogEntry{Term: 1, Data: "test"}, LogEntry{Term: 2, Data: "foo"}}
 
 	test.Expect(server.lastLogTerm()).ToEqual(2)
-}
-
-func TestGenerateRequestVoteDerivedFromLog(t *testing.T) {
-	test := quiz.Test(t)
-
-	server := New()
-	server.Log = []LogEntry{LogEntry{Term: 1, Data: "test"}, LogEntry{Term: 1, Data: "foo"}}
-	newRequestVote := server.RequestVote()
-
-	test.Expect(newRequestVote.Term).ToEqual(1)
-	test.Expect(newRequestVote.CandidateId).ToEqual(server.Id)
-	test.Expect(newRequestVote.LastLogIndex).ToEqual(2)
-	test.Expect(newRequestVote.LastLogTerm).ToEqual(1)
 }
 
 func TestReceiveRequestVoteNotSuccessfulForSmallerTerm(t *testing.T) {
@@ -456,40 +442,6 @@ func TestServerDeletesConflictingEntriesWhenReceivingAppendEntriesMessage(t *tes
 	test.Expect(entry.Data).ToEqual("good")
 }
 
-func TestReceiveVoteResponseEndsElectionForHigherTerm(t *testing.T) {
-	test := quiz.Test(t)
-
-	server := New()
-	server.Term = 0
-	server.State = Candidate
-
-	server.ReceiveVoteResponse(VoteResponseMessage{
-		VoteGranted: false,
-		Term:        2,
-	})
-
-	test.Expect(server.State).ToEqual(Follower)
-	test.Expect(server.Term).ToEqual(2)
-	test.Expect(server.VotesGranted).ToEqual(0)
-}
-
-func TestReceiveVoteResponseTalliesVoteGranted(t *testing.T) {
-	test := quiz.Test(t)
-
-	server := New()
-	server.Term = 0
-	server.State = Candidate
-
-	server.ReceiveVoteResponse(VoteResponseMessage{
-		VoteGranted: true,
-		Term:        0,
-	})
-
-	test.Expect(server.VotesGranted).ToEqual(1)
-	test.Expect(server.State).ToEqual(Candidate)
-	test.Expect(server.Term).ToEqual(0)
-}
-
 func TestServersHavePeers(t *testing.T) {
 	test := quiz.Test(t)
 
@@ -499,113 +451,4 @@ func TestServersHavePeers(t *testing.T) {
 	serverA.AddPeer(serverB)
 
 	test.Expect(serverA.Peers[0]).ToEqual(serverB)
-}
-
-func TestServerCanStartAndWinElectionAfterElectionTimeout(t *testing.T) {
-	test := quiz.Test(t)
-
-	serverA := New()
-	timer := NewElectionTimer(1, serverA)
-	timer.tickerBuilder = FakeTicker
-	defer tiktok.ClearTickers()
-	serverA.ElectionTimer = timer
-	serverB := New()
-	serverC := New()
-	serverA.AddPeer(serverB)
-	serverA.AddPeer(serverC)
-
-	serverA.Start()
-	tiktok.Tick(1)
-
-	timer.ShutDown()
-	test.Expect(serverA.State).ToEqual(Leader)
-	test.Expect(serverA.Term).ToEqual(1)
-	test.Expect(serverA.VotesGranted).ToEqual(2)
-
-	test.Expect(serverB.VotedFor).ToEqual(serverA.Id)
-	test.Expect(serverC.VotedFor).ToEqual(serverA.Id)
-}
-
-func TestServerCanWinElection(t *testing.T) {
-	test := quiz.Test(t)
-
-	serverA := New()
-	serverB := New()
-	serverC := New()
-	serverA.AddPeer(serverB)
-	serverA.AddPeer(serverC)
-
-	serverA.StartElection()
-
-	test.Expect(serverA.State).ToEqual(Leader)
-	test.Expect(serverA.Term).ToEqual(1)
-	test.Expect(serverA.VotesGranted).ToEqual(2)
-
-	test.Expect(serverB.VotedFor).ToEqual(serverA.Id)
-	test.Expect(serverC.VotedFor).ToEqual(serverA.Id)
-}
-
-func TestServerCanLoseElectionForPeerWithHigherTerm(t *testing.T) {
-	test := quiz.Test(t)
-
-	serverA := New()
-	serverB := New()
-	serverC := New()
-	serverA.AddPeer(serverB)
-	serverA.AddPeer(serverC)
-
-	serverB.Term = 2
-
-	serverA.StartElection()
-
-	test.Expect(serverA.State).ToEqual(Follower)
-	test.Expect(serverA.Term).ToEqual(2)
-	test.Expect(serverA.VotesGranted).ToEqual(1)
-
-	test.Expect(serverB.VotedFor).ToEqual("")
-	test.Expect(serverC.VotedFor).ToEqual(serverA.Id)
-}
-
-func TestServerCanLoseElectionDueToOutOfDateLog(t *testing.T) {
-	test := quiz.Test(t)
-
-	serverA := New()
-	serverB := New()
-	serverC := New()
-	serverA.AddPeer(serverB)
-	serverA.AddPeer(serverC)
-
-	serverB.Log = []LogEntry{LogEntry{Term: 1, Data: "some data"}}
-
-	serverA.StartElection()
-
-	test.Expect(serverA.State).ToEqual(Follower)
-	test.Expect(serverA.Term).ToEqual(0)
-	test.Expect(serverA.VotesGranted).ToEqual(1)
-
-	test.Expect(serverB.VotedFor).ToEqual("")
-	test.Expect(serverC.VotedFor).ToEqual(serverA.Id)
-}
-
-func TestServerCanWinElectionWithRetries(t *testing.T) {
-	test := quiz.Test(t)
-
-	serverA := New()
-	serverB := New()
-	serverC := &FailingPeer{
-		numberOfFails: 1,
-		successfulResponse: VoteResponseMessage{
-			Term: 0,
-			VoteGranted: true,
-		},
-	}
-
-	serverA.AddPeer(serverB)
-	serverA.AddPeer(serverC)
-
-	serverA.StartElection()
-
-	test.Expect(serverA.State).ToEqual(Leader)
-	test.Expect(serverA.Term).ToEqual(1)
-	test.Expect(serverA.VotesGranted).ToEqual(2)
 }
